Use web worker to parse fragments & semantics for pairs
