+++ web/tsconfig.json
@@ -19,7 +19,7 @@
     "paths": {
       "@/*": ["src/*"]
     },
-    "lib": ["esnext", "dom", "dom.iterable", "scripthost"],
+    "lib": ["esnext", "dom", "dom.iterable", "scripthost", "webworker"],
     "skipLibCheck": true,
   },
   "include": [
